Don't register MMTk worker threads in runtime thread list
This commit adds a boolean to the Thread::Attach function which controls
whether the calling thread should be added to the runtime thread list.
This is used by the MMTk worker threads in order to not be suspended
during the STW phase. The `ThreadList::SuspendAll()` function will
suspend all threads including runtime and GC threads.

Vanilla ART gets around this as when a GC occurs, ART uses a
`ThreadPool` object to add GC tasks and wake up the "Heap thread pool"
workers so that they can start doing work.

This commit also implements the `stop_all_mutators` and
`resume_mutators` functions.

Change-Id: Ie726936de1951b65fe8f7b43f893dab0695a4e33