Temporarily disable loading boot image for host builds
Loading boot images on host builds results in boot image objects being
added to the root set. Since the binding isn't currently aware of boot
images, the GC fails when trying to copy an object in the boot image. We
temporarily disable loading the boot image until the binding is fixed.

Change-Id: I5a194402d5e4392512a63723b4cd8e80cac5cfec