Don't allow thread suspension before allocation
This commit fixes a bug in the `TryToAllocate` function when
`block_for_gc` is called by MMTk wherein the allocating thread was
suspended and its thread suspension state was not updated. This resulted
in assertion failures when using a debug build. This commit fixes the
bug by maintaining the thread suspension state correctly.

Change-Id: Id4e7c0a8d6e1aaf7c77d5519eaefef5d0d26af9f