Mask MMTk GC state when getting class pointer
MMTk stores GC state (forwarding bits and mark bit) in the two
lowest-order bits of the class pointer in the object header. However, we
want to occasionally read the class pointer while performing a GC (such
as when we want to get the object size), resulting in segfaults due to
a dirty class pointer. We fix this by masking the two lowest-order bits
when getting the class pointer when using MMTk.

Change-Id: I583930ae6abbe94e4d32b9ce26ec6687ef7e51c8