Allow disabling of the write barrier for the SemiSpace GC
This commit removes the write barrier requirement for SemiSpace. It also
guards the use of the card and mod union tables when the write barrier
is disabled.

Note that currently we can only disable the write barrier if we have no
boot images and Zygote since the write barrier tracks pointers from the
Zygote/immune regions into the moving spaces. An alternative with no
write barrier would be to use the immune regions as a root set in order
to find objects which reference objects in moving spaces.

Change-Id: I83da3c7ff851d57dc28eab349f1583a9e88e7fdf