WIP: Allow setting heap size for a particular app
This commit allows us to set the heap size of a targeted app. The heap
size along with various other internal GC triggering heuristics are set.
Currently the targeted application is hardcoded to a test application.

This commit also prints the performance results in the log instead of on
stdout. We hook into ART's SIGQUIT handler for reporting performance
results.

Change-Id: I8a0156bf39654c6896d17fa944d2e0821b24eba2