Always define `Thread::GetMmtkMutator` and cleanup code
Previously `Thread::GetMmtkMutator` was conditionally compiled only when
`ART_USE_MMTK` was set. But this was inconvenient as all code that used
MMTk's mutator would also have to be conditionally compiled. We now
always define the `Thread::GetMmtkMutator` function but return a
`nullptr` if `ART_USE_MMTK` is not set.

Change-Id: I6b60ecc7b9a1cd2180d587e9bf9fca937be79bba