Fix race condition in thread initialization
Vtable for a derived class is not set up until its constructor runs.
Previously we were spawning a new thread inside the constructor of the
base class. If the new thread was scheduled before the constructor of
the derived class then the thread would use the incorrect implementation
of the `MmtkWorkerThread::Run` function. We fix the race by explicitly
creating/starting the pthread inside each constructor.

Change-Id: Id8ff29a19bb0c020c9f11dcc972901ee363bfe0f