Refactor some root visiting sites to avoid reusing slots
ART sometimes reuses slots for different roots which is incompatible
with MMTk's design. We only fix the slot reuse for ClassTable and
DexCache roots in this commit as they were the only ones that were
progression-blockers since these roots are visited during object
scanning/transitive closure time as opposed to at the start.

Change-Id: Ib673326f75372e29248bc8d91b1d3af94f5a39a0