Properly handle non-moving object allocation
This commit also fixes the `Heap::IsMovableObject` and
`Heap::IsObjectLiveLocked` checks. The `Heap::IsMovableObject` check is
required by the `CopyObjectVisitor` used in `Object::Clone`. Previously,
since the check always returned `false`, ART used to allocate the new
objects with the non-moving allocator, quickly filling it up and
resulting in frequent out-of-memory errors.

Change-Id: Icee3f6983a38864befb383578d1ae22d1f02ad39